requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.1.7
django-cors-headers==4.0.0
django-import-export==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
Pillow==9.5.0
PyJWT==2.7.0
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tablib==3.4.0
uritemplate==4.1.1
xlrd==2.0.1
xlwt==1.3.0
